+++ C++/Hollowrectable.cpp
