+++ C++/Palindrome.cpp
