+++ C++/Pattern.cpp
