+++ C++/Program to delete vowels from a given string.cpp
