+++ C++/Program to find the first repeating element in an array of integers in C++.cpp
@@ -1,3 +1,6 @@
+#include<iostream>
+using namespace std;
+
 int main(){
   int array[100], n, i;
   cout << "Enter number of elements: ";
