+++ C++/Pyramid.cpp
