+++ C++/Rhombuspattern.cpp
