+++ C++/butterflypattern.cpp
