+++ C++/invertedhalfpyramid.cpp
