+++ C++/linearsearchinarray.cpp
