+++ C++/pascalstriangle.cpp
