+++ C++/reverse.cpp
