+++ C++/structures.cpp
