+++ C/floyd's_triangle.c
