+++ C/linked list implimentation.c
